teach eltmc_googlesync to inspect committee roles and get email copy fields
Previously, if no-one is assigned to a role, it will generate no email aliases,
and the corresponding google group will be deleted.  This is not what we want, we want
it at least be copied to the group admin user address.

We also want to allow roles to copy email to another address, so that obsolete
roles can be unpublished but remain defined so as the corresponding group
continues to forward mail to another role.

This commit adds a function to perform the appropriate inspection of committe role
to get the additional addresses, and generate a group even if no users are
assigned.